refactor auth module
Restructure module
Separate JWT auth variants
Cleanup code
Fixes #189
